feat: 예약이 존재할 때 예약 삭제 방지 기능 구현

+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -1,8 +1,12 @@
 package roomescape.service;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 import roomescape.controller.dto.CreateReservationTimeRequest;
+import roomescape.dao.ReservationDao;
 import roomescape.dao.ReservationTimeDao;
+import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
 
 import java.util.List;
@@ -11,9 +15,11 @@ import java.util.List;
 public class ReservationTimeService {
 
     private final ReservationTimeDao reservationTimeDao;
+    private final ReservationDao reservationDao;
 
-    public ReservationTimeService(ReservationTimeDao reservationTimeDao) {
+    public ReservationTimeService(ReservationTimeDao reservationTimeDao, ReservationDao reservationDao) {
         this.reservationTimeDao = reservationTimeDao;
+        this.reservationDao = reservationDao;
     }
 
     public List<ReservationTime> readAll() {
@@ -26,6 +32,10 @@ public class ReservationTimeService {
     }
 
     public void deleteTime(int id) {
+        List<Reservation> reservations = reservationDao.findAllByTimeId(id);
+        if (!reservations.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "해당 예약시간에 예약이 등록되어 있습니다.");
+        }
         reservationTimeDao.delete(id);
     }
 }
